CWMSREGI-12998 - Resolving issue with the ScriptableInflowImpl where the ICurrentDayControl used incorrect look forward/back values.  This was causing a discrepency in the saving of forward data for months with different lengths.
Resolving issue in InflowDataAdapter, where it was pulling an extra month of data from the past in some cases.  Also reduced the past days from 2 to 1, so we're not pulling more data than necessary for priming the computations. - Spoke with both Peter and Caleb about this.

Added unit tests for the date function in the InflowDataAdapter, so we can make sure we're computing the expected dates.

Copied from Perforce
 Change: 110572